+++ policyengine_us/variables/household/income/spm_unit/spm_unit_benefits.py
@@ -19,6 +19,7 @@ class spm_unit_benefits(Variable):
             "ak_ssp",
             "dc_ossp",  # DC benefits
             "id_aabd",  # Idaho benefits
+            "ky_ssp",  # Kentucky benefits
             "de_ssp",  # Delaware benefits
             "fl_oss",
             "ks_sspp",  # Kansas benefits
@@ -26,6 +27,7 @@ class spm_unit_benefits(Variable):
             "la_oss",  # Louisiana benefits
             "ma_state_supplement",  # Massachusetts benefits
             "mi_ssp",  # Michigan benefits
+            "me_ssp",  # Maine benefits
             "mn_msa",  # Minnesota benefits
             # California programs.
             "ca_cvrp",  # California Clean Vehicle Rebate Project.